Added option for showing all existing user images
It can be invoked manually through the developer console by calling
showAll(), and it automatically gets invoked once everything finishes
loading if the showAllImages flag declared at the beginning of
party_phaser.js is set.
It is currently turned on, but will be turned off for production.

Also changed SoundManager in effectClient.js to accept sounds as objects
that have a 'location' and a 'loop' properties, in order to support
looping sounds.