Upgraded to Phaser 3.16.1 and made it work
Added a lifetime calculation for SingleAnimatedParticle which kills the
particle when the animation ends.
I still left the code that makes it stay on the last frame when it reaches
it, just in case some race condition or floating point error might make
the particle process another frame after it's supposed to have died.

Also added names to the levels, which aren't used as keys so they
technically *can* be the same, but if there are two levels with the same
name and both have animations, one of them will cause an error because the
animation won't be able to be created.
Come to think of it, adding the index would fix that.
Gonna do that now.
Stop reading this.
It's no longer relevant.
FOR CRYING OUT LOUD, JUST WAIT THREE SECONDS FOR THE NEXT COMMIT!
Sheesh. <_<