Added auto-reattach to EffectClient on reconnections
If the deriving class requests to attach to the server and the server
connection is interrupted, when the server reconnects it will
automatically resend the attach request.
Now about to change ScriptedEffects to use EffectClient so it benefits
from this as well as every other aspect the base class has.