Added reconnection robustness to ScrptedEffects
Now gets the scripts list anew every time it reconnects, and whenever it
does, it removes scritps that were loaded and are not present in the new
lists, loads scripts from the new list that weren't loaded alread, and
leaves the rest untouched.
It makes it so you can change which scripts should show in
Effects/ScriptedEffects/effect.js, restart the server and the clients will
change their contents how you'd expect.