Fix unencountered bug with set volume command
It was sending the entire User object instead of just the username.
It wasn't a problem because this parameter is only meant to be used to
give the user an error message if they use bad syntax, and that feature is
off (and untested) at the moment.
So... fixed that.