Parrot Mate: Text bubbles, image resizes
Added a new TextAction, added to an event by Text('Your text here'), which
shows up when the event is performed in a speech bubble to the parrot's
top-right.
Also resized and moved the images around for a bit so that they are all
the same size and located so that the stand in all of them is in the same
place. Before doing so, when the parrot was animated it looked like the
stand was moving, which was just weird.