Parrot Mate: Extra measures to hide bubble on load
It was showing up on OBS (though not on Chrome), so now I've set it to
have its visibility set to 'hidden' on load and only after using .hide()
with jQuery I set it back to 'visible'.
Also set the default delay to 5 minutes, as it is meant to have in
production.