Added image display script + everything required
Currently works with glow, shadow and dundundun effects.
Partially moved the image display commands from Chatbot.
Added tree structure for pages with EffectClient objects that contain
frames of other EffectClient windows.
Added a blocking action queue functionality to the basic EffectClient
class which allows scripts to queue incoming requests.
This is used for displaying images one after the other.
It is stored in the root of the tree, so it should work across multiple
effects (not tested yet).
This was originally meant for allowing multiple image display effects to
coordinate when images are displayed so they don't overlap, but I ended up
implementing the ImageDisplay effect in such a way that multiple scripts
can connect to it and it shows all of their images.
For that purpose I've added another way to attach to server effects: tags.
This lets one client attach to multiple server effects, which I will use
to let the various image display effects display their images on the
single ImageDisplay client.