Added working support for multi-blocking events
So now you can, for example, block on "Image" and on "Sound" separately,
and when you're done with one it'll proceed to the next in that queue, and
you can also block on both at the same time, and then they both need to
unblock and run before the queues continue.
It works well with imgdisp.js.