Added integration with StreamElements by SEManager
Currently only deals with user points.
There are possible race conditions in all of that - you can add/remove
points concurrently while consuming points and then your test of whether
the user has enough points for something could be messed up.
I don't really know how to deal with this in a simple fashion without
adding delays though, and the risk of it happening or of how much damage
it will cause if so are both not large, so I'm leaving it as is.

Next up I'll add costs to commands, messages from Fluxistant to go with
them and then logs, to have records of what we're doing in case some
trouble does occur.