Added SE web socket integration for events
It is currently unused, but there is now an event notification for each
tip made in SE and for follows to the channel (this event is not broadcast
by tmi.js for some reason).
SEManager also notifies of every cheer, host, raid, sub and store
redemption, but most of these will be taken from tmi.js instead by
TwitchManager, and I'm not sure what to do with store redemptions right
now (we *will* need them later on though, in the Fluxlings project at the
least).

Extracted the event registration and notification system from
TwitchManager into the new EventNotifier class, from which TwitchManager
and SEManager now inherit.