Fixed stuff and added alias support to image commands
Fixed a wrong function name which caused a crash in SEManager.
Changed Assets.getUserFShowerFile() to use lower-case file names and
changed all the file names in the folder to lower case as apparently it
made it not find them when the cases differed.

Immage commands can now specify aliases in commands.json.