Added support for effect-specific command configs
Each script can now have a commands definition file (JSON format) that can
be passed to its inherent commandManager along with a handler for all the
commands in the file (the handler will be given the command as its
argument when invoked).
See ImageCommands and ParrotMate for examples.

Also fixed Parrot Mate to have sequences clear everything themselves
rather than the player doing so - this is so that when sequences are
activated separately, cleanup will still be done.