Added global "reload data" keyboard shortcut
Ctrl+WinKey+F5 now invokes the new reloadData() function in every effect.
Ones that utilize their inherent commandManager to load commands from file
should do so when this function is invokes, in order for it to be possible
to reload these commands after the files have been changed.

By doing this I've discovered a bug where commands would be deemed as
"changed" even though they weren't because the Command object copied data
from the JSON data object in a shallow manner.
This made it so that changes to its .alias[] member were applied to the
original JSON data object too, which then made it seem different than the
same data when it was loaded again.
Solved by using the new Utils.clone function - WHICH, NOTE, CANNOT COPY
EVERYTHING NEATLY!
Only use it for simple objects that contain primitive data and/or arrays
and objects of such simple objects!