Various fixes and minor additions
Blocking events in EffectClient were always supposed to be performed on
the root of the EffectClient tree, but for some reason they weren't, so
they are now.
Removed some old, unused functionality from imgdisp.js.
Added an image to the ImageCommands cache.
Added sound support for text display. It may seem redundant to have yet
another place with that logic, but the sounds need to be synchronized with
the display. For example, if an image is being displayed along with a
sound playing, we want to delay showing text that has a sound until it's
done.