Added BotFight effect, currently works in console
To make it send the messages to the chat instead, switch the lines in the
two say(message) of the bots with the ones currently commented out and
delete the line "debug: true," in the constructor of the BotFight class.

Also added an optional 'enabled' variable to the constructor-given Effect
configuration which allows for completely disabling an effect. This
completely prevents it from loading, though its constructor will still be
invoked and the instance made. So don't put anything with any logic in
there!