Added channel reward redemption notifications
Currently only works for rewards that require text, since those show up as
messages through tmi.js.
To see rewards that don't have extra messages we need to use PubSub, which
is currently not done at all. If I deem it important enough then I'll add
it. I probably will. At some point.

Anyway, now Candy Game recognizes the "Start a Candy Game" reward and
starts a candy game when it's redeemed.
Unfortunately, we cannot automatically accept or reject reward redemptions
at this time - with or without PubSub, as it is not a feature in any of
the Twitch API at this point.
Therefore if the game is already running then the user has lost points for
nothing, which is why we warn about that in the reward description.

If it becomes really important to us, I might consider wrapping a browser
window with a module that would simulate a click of the approve/reject
button in the rewards UI page on Twitch. I'd really, *really* rather avoid
doing that though.