Massive renaming and entry point addition
Up until now the code for each effect was in the file "effect.js" located
in the effect's directory.
This made it so when you have multiple such files open, they're all called
"effect.js" and it's hard to tell them apart (or their path gets added to
the display, which still is inconvenient).
Thus I've decided to change it so that, for example, an effect called
HelloWorld will have two files in its directory: mHelloWorld.js which
holds the actual code, and effect.js, which simply imports
./mHelloWorld.js and exports it again.
The reason I added 'm' at the beginning is to differentiate it from
potential JS files belonging to the effect's web interface. The reason
it's an 'm' of all thing is short for "module," which is the next change I
intend to make - rename "effects" to "modules," since not all of them are
actual effects with web interfaces.