Removed cmdname from aliases for commands.
This is to prevent the cmdname from showing up in the configuration page as an alias.

Also removed all the end-of-line periods in the descriptions of the properties of commands, to fit the style of it everywhere else (tooltips shouldn't have periods).

Lastly, added no-inspection instruction for WebStorm to ignore the unused parameters in the base function invokeCommand(user, command).