Added natural number entities to configurations.
Also fixed a problem with fixed array GUIs during child changes, and removed the validation in ImageCommandEntity that makes sure either an image or a sound is set, since that caused the configuration GUI to give an error whenever a new image command was created.
At worse, if neither is given, it'll be a command that doesn't do anything. *Shrugs*