Added bad configuration rejection on the server side.
Now if the server gets a saveConfig request from a configuration client, if there's a problem with it (i.e. it fails validation) then the new configuration is NOT saved (not even partially), and instead an error event is sent back to the client with the error message along with a description of where it happened (e.g. Modules > Adventure > Minimum Participants).