Uploading and deleting image files works for Image Commands.
No audio or multi-file support yet, and also I need to change the way the files are saved in the user folder. Right now it uses the given filename, but that means that if two different files (different IDs) upload different files with the same name, the last one will override the previous one.
I need to change the data save files so they store both the original filename and the local filename, and have it assign unique filenames for the local files.

Also I need to make it so if a data upload entity is removed, a delete request is sent to the server so as to avoid unused files from littering the server.