Made configurations robust.
When importing configuration files for outdated versions of the bot, it will now import whatever it manages to, discard what it doesn't know and keep the parts it doesn't recognize new information for.
This is to keep users from needing to reconfigure things every time the configuration definition changes for something.