F Shower and Welcome work properly.
Also fixed object titles for user entities so that they show the username when they're in arrays if they have one and the array index if not.
Also removed tabindex from all labels in the configuration page so it doesn't mark them in white.
Also... I have no idea, I've changed and fixed an assortment of things. ^_^;