Added "Advanced Mode" to configuration GUI.
Entities marked with the "advanced" flag will now not show up on the GUI unless the "Advanced Mode" checkbox at the top-right is checked.
To mark an entity as advanced, use .setAdvanced() on it (much like .hide()).

Also extracted the small amount of CSS in the configuration page to its own CSS file in order to facilitate finding a specific style through names rather than indices.