Function Responses: Switched variable replacement to new external mechanism.
Rather than replacing a value in the message and then repeating on the result for all the variables, the new search engine splits the message into parts, replaces the parts that need replacing, then merges it all back together again.
This removes any shred of a chance of user input being mistaken for replacement tokens, since at first only the original message is present, and parts that have been replaced are marked as "final" and will not be replaced with anything again.