Exporting conf files to see the difference in size.
The difference is massive.
Full configuration files can be 40x larger than conf files, and I think that the conf files have all the information needed to define the configuration precisely.

Furthermore, a bigger issue is the size of the HTML page currently generated by the configuration GUI. I got:
    $(document).find('*').length === 13030
Checking online, I found this:
    Lighthouse flags pages with DOM trees that:
    * Have more than 1,500 nodes total.
    * Have a depth greater than 32 nodes.
    * Have a parent node with more than 60 child nodes.
(Taken from: https://stackoverflow.com/questions/34427558/is-too-many-element-in-the-html-can-affect-the-page-performance)

I don't think the configuration page has that last one, but it definitely does the first two.
It shows, too - the page is sluggish even with what it has right now (not that many modules, only default configurations for most things), it takes a while to load and it tends to get stuck after a while which requires reloading it.

I do not have the knowledge and experience I need to create this sort of thing.
I need to find some solution, this isn't viable whatsoever.