Trivia: Ignore punctuation, answer with ban, cats!
Now shows all non-alphanumeric characters in the answer from the start and does not require (but may accept) them in the answer. This has the side effect of accepting punctuation in the middle of answers, but that's an acceptable price.
Also shows the answers to banned questions, to sate the curiosity of certain curious cats.
New command (default !triviacats and !triviacategories) shows the available categories for the trivia game.