Image Commands: Text display supports variables.
Added a `.compileText()` function to the `data` object passed to Function invocations. Invoking it on a string compiles the string like a response text, taking all variables into account.
The text string used for display in the Image Commands module now uses that in order to support variables in the displayed text.