Feeding Game: Add Lag Skip command
When the server sends a feeding event, it waits for a response asynchronously, and nothing cancels that, ever.
If the game client isn't loaded, nothing sends back a response and that event is lost.
Then if the game client is loaded and another feeding event is sent from the server, when a reply is received, it goes towards the previous one, so it processes the one before instead of the new one.
Adding timeouts to these requests is... very problematic, and not worth the effort right now.
So as a workaround, I've added a command (triggered from the CLI by default) to fix this lag by forcing the game client to send a "feedDone" event in order to trigger the current event in the server's queue.