Implement custom filters for triggers
This adds infrastructure to allow specific triggers to provide a list of allowed filters - only those filters will be displayed to the user as options.
Additionally, it makes it possible to easily set filters as a non-advanced for a trigger.
The SR skin purchase triggers currently make use of the latter, and one makes an example use of the former.