Pokyecats: Don't save by default in interface methods
This is to provide a way to perform many actions rapidly without saving after each one, which results in frequent write errors.
This is used only in the Twister module right now, where that issue does indeed happen, and a global save is triggered after everything a single time to save all the changes.