Config: Performance improvement hack
This makes all "folder" GUI elements store their contents inside the actual HTML page only when they're open.
When they close, they save the data aside and remove it from the page.
This doesn't play that well with unsaved changes when switching modules, but I think it's worth it, for now at least.